+++ crates/server-config/src/node_config.rs
@@ -22,6 +22,7 @@ use fs_utils::to_abs_path;
 use particle_protocol::ProtocolConfig;
 use types::peer_id;
 
+use crate::avm_config::AVMConfig;
 use crate::keys::{decode_key, decode_secret_key, load_key};
 use crate::system_services_config::{ServiceKey, SystemServicesConfig};
 use crate::{BootstrapConfig, KademliaConfig};
@@ -84,14 +85,13 @@ pub struct UnresolvedNodeConfig {
     #[serde(default)]
     pub protocol_config: ProtocolConfig,
 
+    /// These are the AquaVM limits that are used by the AquaVM limit check.
+    #[derivative(Debug = "ignore")]
+    pub avm_config: Option<AVMConfig>,
+
     /// Number of AVMs to create. By default, `num_cpus::get() * 2` is used
     #[serde(default = "default_aquavm_pool_size")]
     pub aquavm_pool_size: usize,
-
-    /// Maximum heap size in bytes available for an interpreter instance.
-    #[serde_as(as = "Option<DisplayFromStr>")]
-    #[serde(default)]
-    pub aquavm_max_heap_size: Option<bytesize::ByteSize>,
 
     /// Default heap size in bytes available for a WASM service unless otherwise specified.
     #[serde_as(as = "Option<DisplayFromStr>")]
@@ -196,8 +196,8 @@ impl UnresolvedNodeConfig {
             services_envs: self.services_envs,
             protocol_config: self.protocol_config,
             aquavm_pool_size: self.aquavm_pool_size,
-            aquavm_heap_size_limit: self.aquavm_max_heap_size,
             default_service_memory_limit: self.default_service_memory_limit,
+            avm_config: self.avm_config.unwrap_or_default(),
             kademlia: self.kademlia,
             particle_queue_buffer: self.particle_queue_buffer,
             effects_queue_buffer: self.effects_queue_buffer,
@@ -355,11 +355,11 @@ pub struct NodeConfig {
     /// Number of AVMs to create. By default, `num_cpus::get() * 2` is used
     pub aquavm_pool_size: usize,
 
-    /// Maximum heap size in bytes available for an interpreter instance.
-    pub aquavm_heap_size_limit: Option<bytesize::ByteSize>,
-
     /// Default heap size in bytes available for a WASM service unless otherwise specified.
     pub default_service_memory_limit: Option<bytesize::ByteSize>,
+
+    /// These are the AquaVM limits that are used by the AquaVM limit check.
+    pub avm_config: AVMConfig,
 
     pub kademlia: KademliaConfig,