+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 # Changelog
 
+## [0.22.0](https://github.com/fluencelabs/nox/compare/nox-v0.21.4...nox-v0.22.0) (2024-03-05)
+
+
+### ⚠ BREAKING CHANGES
+
+* **aquavm:** aquavm mem limits from config [fixes VM-425] ([#2111](https://github.com/fluencelabs/nox/issues/2111))
+
+### Features
+
+* **aquavm:** aquavm mem limits from config [fixes VM-425] ([#2111](https://github.com/fluencelabs/nox/issues/2111)) ([a732f5c](https://github.com/fluencelabs/nox/commit/a732f5ca16e2cd5a721a220ce8f73b784dd0ab21))
+
+## [0.21.4](https://github.com/fluencelabs/nox/compare/nox-v0.21.3...nox-v0.21.4) (2024-03-03)
+
+
+### Features
+
+* bump fluence-app-service ([#2130](https://github.com/fluencelabs/nox/issues/2130)) ([98161b1](https://github.com/fluencelabs/nox/commit/98161b1d22ae4d62383c946c5ae415f5e9c5b9ed))
+
 ## [0.21.3](https://github.com/fluencelabs/nox/compare/nox-v0.21.2...nox-v0.21.3) (2024-03-01)